Cut the import time by ~10ms by deferring the import of the `DOMAINS`, `TLDS` and `STOPLIST`. This will add the 10ms to the first call of the method but will not affect subsequent calls. Inspired by https://adamj.eu/tech/2023/03/02/django-profile-and-improve-import-time/.

Main (note that timings are a factor 1000 higher in the display, so 1s on the graph is 1ms in reality):

![image](https://user-images.githubusercontent.com/12661555/223687935-b4305e94-0dfa-4901-8a6c-ad4fda539b8a.png)

This branch:

![image](https://user-images.githubusercontent.com/12661555/223687797-7a7dcd5d-aba9-4a0c-bd7c-74b7bb68c849.png)

Runtime unaffected:

```
pyswot version: 0.3.0
10000 loops for user@gmcc.tnua.edu.tw, best of 5: 6.785823900027026e-06 s
10000 loops for user@nope.tnua.edu.tw, best of 5: 5.781288719990698e-06 s
```